user deletion for admin menu working, no duplicates in adding storages allowed

+++ src/main/java/com/codecool/dao/database/DatabaseUserDao.java
@@ -72,7 +72,7 @@ public final class DatabaseUserDao extends AbstractDao implements UserDao {
 
     @Override
     public void delete(int userId) throws SQLException {
-        String sql = "DELETE FROM users WHERE id = ?;";
+        String sql = "DELETE FROM users WHERE id = ? AND admin != true;";
 
         try (PreparedStatement statement = connection.prepareStatement(sql)) {
             statement.setInt(1, userId);
